pyproject.toml
[tool.poetry]
name = "fqlearn"
version = "0.1.0"
description = ""
authors = ["EverVino <evervino00@gmail.com>", "Faith Hunja <faithhunja17@gmail.com>"]
readme = "README.md"
packages = [{include = "fqlearn", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.8.2"
pandas = "^2.1.3"
numpy = "^1.26.2"
scipy = "^1.11.4"

#[tool.pytest.ini_options]
#python = ["src"]
python-ternary = "^1.0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
